Exchanged names of mixed files
FlatMap and TimestampTimeInterval were mixed
